SIGAA Upgrade
Now i implemented my version of the site. This versions is a way better than the original im my conception because fits on every single screen size and have several design improvements.
This is my first "big" broject and i spend a few hours here. A friend tell me to add a dark mode and maybe i will make it soon, another ideia is to code in this desing style the main page of SIGAA, the screen that appear right after you login and shows you your classes, activities, some menus with many other pages like your classmates, student contracts and more, and i think will be a great challenge.

PS: I am thinking if i will contact the university to show then the website and maibe get a job or some money for that.